# frontend/src/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: TinBoker Backend API
  description: Backend API for TinBoker - Financial podcast insights platform
  version: 1.0.0
servers:
- url: https://api.tinboker.com
  description: Production server
- url: https://staging-api.tinboker.com
  description: Staging server
- url: https://dev-api.tinboker.com
  description: Development server
- url: http://localhost:8000
  description: Local development
paths:
...
  /api/content/{ticker}:
    get:
      tags:
      - content
      summary: Get Ticker Content
      operationId: get_ticker_content_api_content__ticker__get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          title: Ticker
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/podcast:
    get:
      tags:
      - podcast
      summary: Get Sorted Podcasts
      description: 'Get sorted podcasts list


        Query params:

        - sort_by: Sort field (name, episode_count, created_at, updated_at)

        - order: Sort order (asc, desc)

        - limit: Maximum number of podcasts to return (1-200)

        - offset: Pagination offset


        CDN Cache: 30 minutes'
      operationId: get_sorted_podcasts_api_podcast_get
      parameters:
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          description: Sort field (name, episode_count, created_at, updated_at)
          default: name
          title: Sort By
        description: Sort field (name, episode_count, created_at, updated_at)
      - name: order
        in: query
        required: false
        schema:
          type: string
          description: Sort order (asc, desc)
          default: asc
          title: Order
        description: Sort order (asc, desc)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of podcasts to return (1-200)
          default: 50
          title: Limit
        description: Maximum number of podcasts to return (1-200)
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Podcast'
                title: Response Get Sorted Podcasts Api Podcast Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/podcast/{podcast_name}:
    get:
      tags:
      - podcast
      summary: Get Podcast By Name
      description: 'Get podcast by name


        Returns podcast metadata including episode count and timestamps


        CDN Cache: 30 minutes'
      operationId: get_podcast_by_name_api_podcast__podcast_name__get
      parameters:
      - name: podcast_name
        in: path
        required: true
        schema:
          type: string
          description: Podcast name
          title: Podcast Name
        description: Podcast name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Podcast'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/podcast/{podcast_name}/episodes:
    get:
      tags:
      - podcast
      summary: Get Podcast Episodes
      description: 'Get episodes for a podcast


        Query params:

        - sort_by: Sort field (created_time, episode_number, episode_title)

        - order: Sort order (asc, desc)

        - limit: Maximum number of episodes to return (1-200)

        - offset: Pagination offset


        CDN Cache: 30 minutes'
      operationId: get_podcast_episodes_api_podcast__podcast_name__episodes_get
      parameters:
      - name: podcast_name
        in: path
        required: true
        schema:
          type: string
          description: Podcast name
          title: Podcast Name
        description: Podcast name
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          description: Sort field (created_time, episode_number, episode_title)
          default: created_time
          title: Sort By
        description: Sort field (created_time, episode_number, episode_title)
      - name: order
        in: query
        required: false
        schema:
          type: string
          description: Sort order (asc, desc)
          default: desc
          title: Order
        description: Sort order (asc, desc)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of episodes to return (1-200)
          default: 50
          title: Limit
        description: Maximum number of episodes to return (1-200)
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      - name: include_content
        in: query
        required: false
        schema:
          type: boolean
          description: Include heavy content fields (transcript, summary)
          default: false
          title: Include Content
        description: Include heavy content fields (transcript, summary)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Episode'
                title: Response Get Podcast Episodes Api Podcast  Podcast Name  Episodes
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/episodes/recent:
    get:
      tags:
      - episodes
      summary: Get Recent Episodes
      description: 'Get recent episodes across all podcasts, sorted by created_time
        descending


        Query params:

        - limit: Maximum number of episodes to return (default: 20, max: 200)

        - offset: Pagination offset

        - podcast_name: Optional filter by podcast name


        CDN Cache: 30 minutes'
      operationId: get_recent_episodes_api_episodes_recent_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of episodes to return (1-200)
          default: 20
          title: Limit
        description: Maximum number of episodes to return (1-200)
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      - name: podcast_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional filter by podcast name
          title: Podcast Name
        description: Optional filter by podcast name
      - name: include_content
        in: query
        required: false
        schema:
          type: boolean
          description: Include heavy content fields (transcript, summary)
          default: false
          title: Include Content
        description: Include heavy content fields (transcript, summary)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/episodes/by-ticker/{ticker}:
    get:
      tags:
      - episodes
      summary: Get Episodes By Ticker
      description: 'Get episodes that mention a specific stock ticker


        Path params:

        - ticker: Stock ticker symbol (e.g., "NVDA", "AAPL")


        Query params:

        - limit: Maximum number of episodes to return (default: 50, max: 200)

        - offset: Pagination offset


        CDN Cache: 30 minutes'
      operationId: get_episodes_by_ticker_api_episodes_by_ticker__ticker__get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          description: Stock ticker symbol
          title: Ticker
        description: Stock ticker symbol
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of episodes to return (1-200)
          default: 50
          title: Limit
        description: Maximum number of episodes to return (1-200)
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      - name: include_content
        in: query
        required: false
        schema:
          type: boolean
          description: Include heavy content fields (transcript, summary)
          default: false
          title: Include Content
        description: Include heavy content fields (transcript, summary)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/episodes/by-tag/{tag}:
    get:
      tags:
      - mock-data
      summary: Get Episodes By Tag
      description: 'Get episodes with a specific tag


        Returns real episode data from Firestore subcollections filtered by tag'
      operationId: get_episodes_by_tag_api_episodes_by_tag__tag__get
      parameters:
      - name: tag
        in: path
        required: true
        schema:
          type: string
          description: Tag name or ID
          title: Tag
        description: Tag name or ID
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of episodes to return
          default: 50
          title: Limit
        description: Maximum number of episodes to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      - name: include_content
        in: query
        required: false
        schema:
          type: boolean
          description: Include heavy content fields (transcript, summary)
          default: false
          title: Include Content
        description: Include heavy content fields (transcript, summary)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpisodesByTagResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EpisodesByTagResponse:
      properties:
        tag:
          type: string
          title: Tag
        episodes:
          items:
            type: object
          type: array
          title: Episodes
        total:
          type: integer
          title: Total
      type: object
      required:
      - tag
      - episodes
      - total
      title: EpisodesByTagResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Podcast:
      properties:
        id:
          type: string
          title: Id
          description: Podcast identifier (podcast_name)
        name:
          type: string
          title: Name
          description: Podcast name
        episode_count:
          type: integer
          title: Episode Count
          description: Total number of episodes
          default: 0
        created_at:
          anyOf:
          - type: integer
          - type: 'null'
          title: Created At
          description: First episode creation timestamp (Unix milliseconds)
        updated_at:
          anyOf:
          - type: integer
          - type: 'null'
          title: Updated At
          description: Most recent episode timestamp (Unix milliseconds)
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
          description: Podcast cover image URL (from most recent episode)
      type: object
      required:
      - id
      - name
      title: Podcast
      description: Podcast metadata model (aggregated from episodes)
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer